samples/cea-techpulse-python/src/requirements.txt
python-dotenv
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal
openai
mcp
gunicorn